+++ app/Http/Controllers/GaleriController.php
@@ -5,11 +5,18 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Galeri;
 use App\Helpers\ImageHelper;
+use App\Traits\HasPermissionCheck;
 
 class GaleriController extends Controller
 {
+    use HasPermissionCheck;
+
     public function index(Request $request)
     {
+        if ($response = $this->checkPermission('galeri.view', 'Anda tidak memiliki izin untuk melihat galeri.')) {
+            return $response;
+        }
+
         $query = Galeri::query();
 
         if ($request->filled('tipe')) {
@@ -34,11 +41,19 @@ class GaleriController extends Controller
 
     public function create()
     {
+        if ($response = $this->checkPermission('galeri.create', 'Anda tidak memiliki izin untuk upload galeri.')) {
+            return $response;
+        }
+
         return view('backend.v_galeri.create', ['judul' => 'Upload Foto Galeri']);
     }
 
     public function store(Request $request)
     {
+        if ($response = $this->checkPermission('galeri.create', 'Anda tidak memiliki izin untuk upload galeri.')) {
+            return $response;
+        }
+
         $request->validate([
             'judul'      => 'required|string|max:100',
             'foto'       => 'required|image|mimes:jpg,jpeg,png|max:3072',
@@ -63,6 +78,10 @@ class GaleriController extends Controller
 
     public function destroy(Galeri $galeri)
     {
+        if ($response = $this->checkPermission('galeri.delete', 'Anda tidak memiliki izin untuk menghapus galeri.')) {
+            return $response;
+        }
+
         ImageHelper::deleteImage($galeri->foto, 'img-galeri');
         $galeri->delete();
